app/backend/requirements.txt
aiobotocore==1.2.0
aiohttp==3.7.2
aioitertools==0.7.0
async-timeout==3.0.1
attrs==20.2.0
botocore==1.19.52
boto3==1.16
botostubs==0.12.1.16.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.9
fsspec==0.8.4
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
multidict==5.0.0
numpy==1.19.3
opt-einsum==3.3.0
pandas==1.1.3
Pillow==8.0.1
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl==1.5.0
pyspark==3.0.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3fs==0.5.1
s3transfer==0.3.3
scikit-learn==0.24.1 
six==1.15.0
torch==1.9.0
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.2
